requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
python-dotenv>=1.0
boto3>=1.34
pytest>=8.2
pytest-django>=4.8

# Optional (MySQL for production; install when targeting MySQL)
# mysqlclient>=2.2

